feat: smarter AI pathfinding/voting + touch-follow controls
AI: stuck detection (re-routes after 1s), 10 extra waypoints for
obstacle avoidance, suspicion-based voting (seen-near-body, unseen
players), contextual chat (accuse/defend by role), up to 3 msgs/meeting.
Fix time.Duration bug with float64 settings (meeting/cooldown timers).

Client: finger-follow touch mode as joystick alternative — toggle in
settings menu, persisted in localStorage. Touch canvas to move toward
finger, lift to stop. HUD buttons don't trigger movement.

+++ server/client.go
@@ -945,11 +945,11 @@ func runMeetingTimeline(room *Room, meeting *Meeting, discTime, voteTime float64
 	}()
 
 	// Discussion phase
-	time.Sleep(time.Duration(discTime) * time.Second)
+	time.Sleep(time.Duration(discTime * float64(time.Second)))
 	meeting.SetPhase("voting")
 
 	// Voting phase
-	time.Sleep(time.Duration(voteTime) * time.Second)
+	time.Sleep(time.Duration(voteTime * float64(time.Second)))
 
 	// Tally
 	ejectedID := meeting.TallyVotes()
